pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "pymupdf4llm>=0.0.17",
    "pdfplumber>=0.11",
    "pytesseract>=0.3.13",
    "Pillow>=10.0",
    "llama-index-core>=0.11",
    "llama-index-embeddings-ollama>=0.3",
    "llama-index-vector-stores-qdrant>=0.3",
    "qdrant-client>=1.9",
    "python-dotenv>=1.0",
    "loguru>=0.7",
    "typer>=0.12",
    "rank-bm25>=0.2",
    "sentence-transformers>=3.0",
    "fastapi>=0.111",
    "uvicorn>=0.30",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff", "mypy"]